+++ src/security/models/user.py
@@ -2,7 +2,7 @@ import uuid
 
 from sqlalchemy import Column, String, Boolean, DateTime as DateTimeA, func
 
-from security.schemas.runtime.guid import GUID
+from security.schemas.guid import GUID
 from helpers.database.database import Base
